\name{mycomb}
\alias{mycomb}
\title{ My Combined Classifier }
\description{
  The list of classifiers to be combined.
}
\usage{
mycomb()
}
\details{
 Stabilised linear discriminant analysis \code{\link{slda}}, k-nearest
neighbours (k=5 and k=10) and the logistic regression model (multinomial
model respectively) are combined.
}
\author{ Torsten.Hothorn <Torsten.Hothorn@rzmail.uni-erlangen.de> }
\keyword{misc}
