\name{print.cvclass}
\alias{print.cvclass}
\alias{print.cvreg}
\alias{print.cvsurv}
\alias{print.bootestclass}
\alias{print.bootestreg}
\alias{print.bootestsurv}
\title{Print Method for Error Rate Estimators}
\description{
 Print objects returned by \code{\link{errorest}} in nice layout.
}
\usage{
\method{print}{cvclass}(x, digits=4, ...)
}

\arguments{
  \item{x}{an object returned by \code{\link{errorest}}.}
  \item{digits}{how many digits should be printed.}
  \item{\dots}{further arguments to be passed to or from methods.}
}
\value{
  none
}
\keyword{misc}
