\name{print.errorest}
\alias{print.errorest}
\title{Print Method for Errorest Object}
\description{
 Print object of class \code{errorest} in nice layout.
}
\usage{
print.errorest(x, digits=4, ...)
}

\arguments{
  \item{x}{object of class \code{errorest}.}
  \item{digits}{how many digits should be printed.}
  \item{\dots}{further arguments to be passed to or from methods.}
}
\details{
 A \code{errorest} object is a named list of numbers and character strings.
}
\value{
  none
}
\author{ Torsten.Hothorn <Torsten.Hothorn@rzmail.uni-erlangen.de> }
\keyword{misc}
