\name{print.inbagg}
\alias{print.inbagg}
\title{Print Method for Inbagg Object}
\description{
Print object of class \code{inbagg} in nice layout.
}
\usage{
\method{print}{inbagg}(x, ...)
}
\arguments{
  \item{x}{object of class \code{inbagg}.}
  \item{\dots}{additional arguments.}
}
\details{
An object of class \code{inbagg} is printed. Information about number and names of the intermediate variables,
and the number of drawn bootstrap samples is given.
}
\keyword{misc}
