\name{print.inclass}
\alias{print.inclass}
\title{Print Method for Inclass Object}
\description{
Print object of class \code{inclass} in nice layout.
}
\usage{
\method{print}{inclass}(x, ...)
}
\arguments{
  \item{x}{object of class \code{inclass}.}
  \item{\dots}{additional arguments.}
}
\details{
An object of class \code{inclass} is printed. Information about number and names of the intermediate variables, the used modelling technique and the number of
drawn bootstrap samples is given.
}
\keyword{misc}
