\name{Smoking}
\alias{Smoking}
\non_function{}
\title{Smoking Styles}
\usage{data("Smoking")}
\description{
The \code{Smoking} data frame has 55 rows and 9 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{NR}{numeric, patient number.}
    \item{Sex}{factor, sex of patient.}
    \item{Age}{factor, age group of patient, grouping consisting of those in their twenties, those in their thirties and so on.}
    \item{TarY}{numeric, tar yields of the cigarettes.}
    \item{NicY}{numeric, nicotine yields of the cigarettes.}
    \item{COY}{numeric, carbon monoxide (CO) yield of the cigarettes.}
    \item{TVPS}{numeric, total volume puffed smoke.}
    \item{BPNL}{numeric, blood plasma nicotine level.}
    \item{COHB}{numeric, carboxyhaemoglobin level, i.e. amount of CO absorbed by the blood stream.}
  }
}

\details{
 The data describes different smoking habits of probands.
}

\source{
 Hand and Taylor (1987), Study F \emph{Smoking Styles}.
}

\references{
D.J. Hand and C.C. Taylor (1987), 
\emph{Multivariate analysis of variance and repeated measures.} London: Chapman \&
Hall, pp. 167--181.
}

\keyword{datasets}
