\name{summary.bagging}
\alias{summary.bagging}
\alias{print.summary.bagging}
\title{Summarising Bagging}
\description{
  \code{summary} method for class \code{bagging}.
}
\usage{
\method{summary}{bagging}(object, \dots)
}
\arguments{
  \item{object}{Object of class \code{"bagging"}.}
  \item{\dots}{further arguments to be passed to or from methods.}  
}
\details{
  A representation of all trees in the object is printed.
}
\value{
  none
}
}
\author{ Torsten.Hothorn <Torsten.Hothorn@rzmail.uni-erlangen.de> }
\keyword{tree}
