\name{summary.inbagg}
\alias{summary.inbagg}
\alias{print.summary.inbagg}
\title{Summarising Inbagg}
\description{
Summary of inbagg is returned.
}
\usage{
\method{summary}{inbagg}(object, ...)
}
\arguments{
  \item{object}{an object of class \code{inbagg}.}
  \item{\dots}{additional arguments.}
}
\details{
A representation of an indirect bagging model 
(the intermediates variables, the number of bootstrap samples, the trees) is printed.
}
\value{
none
}




\seealso{\code{\link{print.summary.inbagg}}}


\keyword{misc}

