\name{summary.inclass}
\alias{summary.inclass}
\alias{print.summary.inclass}
\title{Summarising Inclass}
\description{
Summary of inclass is returned.
}
\usage{
\method{summary}{inclass}(object, ...)
}
\arguments{
  \item{object}{an object of class \code{inclass}.}
  \item{\dots}{additional arguments.}
}
\details{
A representation of an indirect classification model 
(the intermediates variables, which modelling technique is used and the
prediction model) is printed.
}
\value{
none
}


\author{Andrea Peters <Peters.Andrea@imbe.imed.uni-erlangen.de>}


\seealso{\code{\link{print.summary.inclass}}}


\keyword{misc}

