\name{print.classbagg}
\alias{print}
\alias{print.classbagg}
\alias{print.regbagg}
\alias{print.survbagg}
\title{Print Method for Bagging Trees}
\description{
  Print objects returned by \code{\link{bagging}} in nice layout.
}
\usage{
\method{print}{classbagg}(x, digits, \dots)
}
\arguments{
  \item{x}{object returned by \code{\link{bagging}}.}
  \item{digits}{how many digits should be printed.}
  \item{\dots}{further arguments to be passed to or from methods.}  
}
\value{
  none
}
\author{ Torsten.Hothorn <Torsten.Hothorn@rzmail.uni-erlangen.de> }
\keyword{tree}
