\name{smoking}
\alias{smoking}
\non_function{}
\title{Smoking Styles}
\usage{data(smoking)}
\description{
The \code{smoking} data frame has 55 rows and 9 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{NR}{numeric. Patient number.}
    \item{Sex}{factor. Sex of patient.}
    \item{Age}{factor. Age group of patient, grouping consisting of those in their twenties, those in their thirties and so on.}
    \item{TarY}{numeric. Tar yields of the cigarettes.}
    \item{NicY}{numeric. Nicotine yields of the cigarettes.}
    \item{COY}{numeric. Carbon monoxide (CO) yield of the cigarettes.}
    \item{TVPS}{numeric. Total volume puffed smoke.}
    \item{BPNL}{numeric. Blood plasma nicotine level.}
    \item{COHB}{numeric. Carboxyhaemoglobin level, i.e. amount of CO absorbed by the blood stream.}
  }
}

\details{
 The data describes different smoking habits of probands.
}

\source{
 Hand and Taylor (1987), Study F \emph{Smoking Styles}.
}

\references{
D.J. Hand and C.C. Taylor (1987), 
\emph{Multivariate analysis of variance and repeated measures.} London: Chapman \&
Hall, pp. 167--181.
}

\keyword{datasets}
